requirements.txt
loguru
bs4
lxml
pandas
sqlalchemy
rich
feedparser
apprise
python-dotenv
loguru